@@ -209,16 +205,15 @@
 		"vue-cropperjs": "2.2.0",
 		"vue-js-modal": "1.3.13",
 		"vue-json-tree-view": "2.1.4",
-		"vue-loader": "15.0.3",
+		"vue-loader": "15.0.11",
 		"vue-router": "3.0.1",
 		"vue-template-compiler": "2.5.16",
 		"vuedraggable": "2.16.0",
 		"vuex": "3.0.1",
-		"web-push": "3.3.0",
+		"web-push": "3.3.1",
 		"webfinger.js": "2.6.6",
-		"webpack": "4.6.0",
-		"webpack-cli": "2.0.15",
-		"webpack-replace-loader": "1.3.0",
+		"webpack": "4.8.3",
+		"webpack-cli": "2.1.3",
 		"websocket": "1.0.26",
 		"ws": "5.1.1",
 		"xev": "2.0.0"
